requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pdf2image
pathlib
youtube_transcript_api
plotly
pandas
numpy
pydeck
scikit-learn
wordcloud
matplotlib